neubot/privacy.py: Update to privacy policy 2.0
Replace consistently 'can_share' with 'can_publish' and adapt
the logic to cope with the new regime.  The only exception for
now is collect_allowed(), which tells whether we can save the
user's Internet address.  I cannot change this function now for
compatibility with old clients, which continue to run tests w/
the master server.  It will go away in a few cycles.

Introduce a new function, count_valid(), which is the workhorse
for the privacy module, since all other functions are implemented
in terms of it.

Changes tested using the attached regression test.