neubot/privacy: Add command line option to test settings
This commit adds neubot/privacy.py the -t command line option,
that allows to test the current database and fail if permissions
are not set.  This commit will be needed by a piece of code in
the Debian package that needs to prompt the user if permissions
are not OK.

Tested using the regression tests attached to this commit.