neubot/config.py: Update to privacy policy 2.0
This commit modifies the name of the variable used throughout
neubot to indicate the possibility of republishing the raw results
on the WWW.  The name is now ``can_publish``, which is more clear
than ``can_share``.

This commit stands by itself, but I don't guarantee that the tree
will work, until all modules have been adapted to cope with this
new change.