Sanitize certfile before passing it to wrap_socket()
The problem is that wrap_socket() is very picky and considers the
empty string a certificate, so it raises an exception because a key
was not provided.

If look at the code, this works when "net.stream.certfile" is not
defined, because the get() returns None as the default value.  Alas,
configuration variables are always declared now, the declared value
for "net.stream.certfile" is "" and breaks wrap_socket().

Verified that now we have fixed the problem with the http.client
subcommand and an https: URI.