net/stream.py: return to settings for localhost
Some six months ago I was planning to make stream.py a real test,
but eventually that didn't happened.  Moreover, neubot.blupixel.net
is not running the stream test anymore.  So, get back to settings
optimized for localhost, since the only reason to use this file is
for measuring performances.

Verified that everything works after this change by running the
client and server of stream.py.