agent.py: net.stream does not provide default settings anymore
We shall not pass object.configure() a dictionary with just a
subset of the settings because portions of the code don't provide
defaults anymore.  This is being done to reduce confusion and
bugs and ensure things are defined, if possible, in one single
place.

For example, recently the net.stream module has been changed to
stop providing default settings.  And it turns out that this
piece of agent.py was instead assuming that stream would have
supplied default settings.  Fix by adding settings to the current
configuration and passing it to configure().

I have also checked with ``git grep -C15 configure neubot``
that there are no other occurrences of this bug.

Of the three settings added to the dictionary, if rootdir or ssi
are wrongly passed, things does not work and I will notice that
for sure.  The third setting, bind_or_die, is less obvious and so
I've made sure that, if we pass a bad port, the program still
fails to start up thanks to that setting.