log.py: delayed string interpolation
Prodded by pylint I noticed that I should delay string interpolation
in log.py until I know that a string will be actually printed, as the
logging module WISELY already does.

The thruth is that it was a mistake to drop the logging module
frontend, because I was not happy with the backend, and, especially
with the syslog backend.

This commit goes in the right direction because it adds a simple API
more compatible with logging.

With a bit of time, I'll be able to cook a mechanical diff that
replaces LOG with logging.

Once I've done that, the next step will be to use the logging frontend
and my current code as the backend.