MacOS: Update PrivacyPane to privacy policy 2.0
This commit updates the PrivacyPane to privacy policy 2.0,
which means that I've replaced all the instances of can_share
with can_publish and I've modified the algorithm so that you
cannot install Neubot unless you assert you're informed and
you provide the permission to collect and share.

I am committing this part alone because it is a selfcontained
piece of code and I was able to check that everything works
by running the installer and making sure that you are allowed
to continue with the installation only when all permissions
are given.  Moreover I've checked that you can go back to the
previous pane regardless of the privacy settings, which was
also the past behavior.