log.py: Introduce streaming API
Allows to copy logs to a set of streams.  Needed to pass logs to the
client when the test is delegated to the local daemon using runner API
(which will be added in the next commits).

I needed to push lazy processing down the stack because log streaming
must be in a vantage point that allows to see all the logs: now lazy
processing is performed before streaming, if streaming is on, and after
the check for verbosity otherwise.  Of course, when we are streaming
and we are verbose it is attempted just once.