runner_api.py: NOT (streaming AND overlapped tests)
If the user requests for log streaming, do not allow to start a new
test if an existing test is already in progress.  Given the way I have
designed log streaming, what the user will see in this case may be a
bit too confusing.  Indeed, log streaming will start immediately, and
so the user will see the in progress test.  Then, the log streaming
will be closed when the requested test is about to start.

I don't see a real use case for allowing overlapping tests from the
command line, so I think it's simpler to add this check than to adapt
the API to cope with overlapping tests and streaming.