runner_clnt: select return value more carefully
Naively we returned False on error, which means you run a test
on your context if a test is already running on the local daemon,
and that sucks.  Better to return False if the local daemon is
not reachable or communication with it seems unfeasible and True
otherwise, regardless of errors (which may indeed be because a
test is already running).