Rewrite /api/config and zap /api/configlables
I want to add some "post config" hooks, while there I take the
opportunity to move the code into another file and refactor such
that /api/configlabels is no longer needed.

I will delay the commit of the post config hooks for a while,
because the patch is not ready, but this piece can for sure go
in for testing.

Tested by making sure that query string options are honored, that
invalid privacy settings and invalid interval settings are detected,
and that, more generally, the web interface seems OK.