runner_core: let run() select negotiate URI
More abstraction, let run() select negotiate URI instead of providing
it the URI.

Will allow to repair (invoking the rendezvous) when the list of registered
tests is empty.

Also, update the unit test and split the bittorrent case in two: one is
when there is a negotiate URI, the other when there isn't.

Verified that this works with and without negotiate URI using the internal
main() for testing and the updated unit test.

Another patch that updates the remainder of the code to invoke run() with
two instead of three arguments is due.