runner_clnt: print the full exception
I've noticed that some httplib exceptions don't have an attached text and
the resulting error message is empty.

Better to print the full stack trace, may be confusing for users but is
what I need to see when they ask for help.

Verified this works running the internal main() when the local daemon isn't
running.