runner_core: do not stop streaming immediately
When asked to run a nonexistent test, "testdone" is published immediately
and log streaming is stopped before the HTTP response is sent.

As a result, httplib raises and error and complains that the HTTP response
was empty, masking the real error.

Repair by stopping streaming a couple of seconds after the test is done
so there is time to send the response.

Verified this works by running a bittorrent test and a foobar test and
seeing what happens on console.