utils_path.py: util functions for paths
This is much more general than what is needed to implement the
file-based backend.  I've studied again the whole path processing
and string encoding stuff and come out with better code.

Garnered confidence that this works with the attached unit test,
discussed in detail with Paola and Giuseppe.

The patch also looks good to pylint and python2.7-2to3.