filesys: file system abstraction
Minimal filesystem abstraction needed to implement the new
backend infrastructure.  The mid-term plan is to move there
all filesystem code and to zap the user-wide database in the
process.

Another noteworthy innovation of this commit is that datadir
changes depending on whether the OS is Linux, or another UNIX.
In the former case, the choice is /var/lib/neubot, in order
to be compliant with FHS.  Otherwise it is /var/neubot as it
always used to be.

Tested using the unit test and, together with the backend
code, on our main server for a couple of days.

The code looks also good to pylint and python2.7-2to3.