backend: generic backend for saving results
Generic backend for saving results, so that we can switch transparently
between different real backends.

The 'neubot' backend saves the result into an sqlite3 database and
is typically used on the user side.

The 'mlab' backend follows Measurement Lab conventions and each test
is a compressed JSON file in a given directory.

Prototype code neubot/backend.py discussed with and written by Fabio
Forno, thanks!

Has been running for a number of days on the Neubot master server
without problems, and pylint and 2to3 are happy too.