utils_posix.py: pass uid, gid to mkdir and touch
Be prepared for when we will have chroot and pass to mkdir and
touch the numeric identifiers not the user name.

Requires some mechanical changes in the unit test.  Plus, I have
also omitted, in the unit test, the user and group identifier when
they are redundant.

No pylint or 2to3 changes after this commit.