Clear previous test results when we change test name
We used to clear previous tests results at the beginning of the
test phase, but previous test name is cleared at the beginning of
negotiation instead.

So, if a test negotiates but fails to run, the status bar will
show the new test name with the old test results.

Fix by clearing previous test results when we change the test name,
at the beginning of negotiation.

Tested by running a manual test, with eth0 down, after another
succesful manual test and verifying that the previous test results
are indeed cleared.

Reported by Alessio Palmero Aprosio, thanks!