scripts/collect.sh: scalable data collection
This patch adds to the tree the script I'm working on to implement
scalable data collection.

The part that pulls results from the remote server is OK, what is
missing is the postprocessing.

We need a bit of postprocessing because results with all privacy
settings can be published immediately, while other results need
to be anonymized first.