M-Lab: fail deploy if we cannot stop a running Neubot
The command used to stop a running Neubot was hiding errors because
it wanted to catch the case where the stop file does not exist because
we've not deployed yet.

Unfortunately, this also catches other type of errors, e.g. that the
pidfile does not contain the right PID.

Repair by checking in advance that the stop file exists and, if so,
run it and do not catch any error.

Fixes the deployment problem seen with the mlab3.lax01 host.

Tested by checking that the deployment fails if the pidfile is wrong
and succeeds if the pidfile is OK.