Check privacy when running command line tests in the local context
Privacy settings should be checked also when you run command line tests
in the local context.

This is not a real issue because this has always been the developers' way
of invoking tests.  And is even less critical now that by default tests
are run via the runner API in the context of the server.

Still, this is a bug and must be fixed.

Noticed when helping Tiziana to understand why automatic tests do not
run on her laptop.

Tested by running speedtest and bittorrent from command line in a virtual
machine where Neubot is not installed.  Verified that it complains when
there are no privacy settings and that it works if I override the settings
via command line.