scripts: extract code to update_changelog from release
Now I don't guarantee anymore that the history is linear and the
immediate effect is that it's not guaranteed that 0.4.8-rc1 is an
ancestor of 0.4.8-rc2.  The only guarantee is that a stable release
will be an ancestor of the next stable release.  The consequence
is that I cannot write the ChangeLog for release candidates anymore.

Hence, the decision to separate the code to update the changelog
and the code to bless a release.