MacOS/makepkg.py: do not create the latest file
This file is dangerous in the sense that if I'm tired and make a
new MacOSX testing release and update all the dist directory by
mistake I can force an update for a number of users that are instead
following the stable channel.  It is much more robust, against my
possible errors, that I have to vi into the already existing testing
file and modify its content.