browser: open page in the browser
Add simple module to open page in the browser, which is intended to
wrap and/or complement webbrowser.

The generic module makes sure that the URI looks like an HTML page
for additional robustness.

The Win32 module bypasses webbrowser and invokes os.startfile() because
what we need is to be sure it does not block, not flexibility.

The null module pretends to open a web page, but actually just prints
a message using the logger.

Tested making sure that browser.py refuses to open things that don't
look like an HTML page URI and that, on Windows, Neubot web pages and
other pages are correctly opened.

Also, the commit looks clean to pylint.