notifier_browser: configurabile browser-based notifier
This is a notifier that uses the browser.py module backend and allows
the user to configure some settings.

The user can now tell the notifier not to open the browser when Neubot
is disabled.  This is not the default behavior because I believe
there is a value in being notified, but this has been requested by
some folks, so I'm adding it.

Also, the user can tell the notifier the minimum interval between two
consecutive notifications for the same URI.  The default value for
this setting is 24 hours. I.e. you get one notification at startup
and then, if you don't shutdown your computer, the next notification
is after 24 hours.

Tested from the command line on windows XP and 7, making sure that
we honor enabled and min_interval, and that default settings prevent
opening multiple notifications.

Also, the commit looks clean to pylint.