runner_core.py: avoid two consecutive rendezvous
When the test name is rendezvous and the test list is empty the code
will schedule two consecutive rendezvous.  We don't want that to happen,
of course.

Verified invoking runner_core.py from the command line that, after this
change, when the test name is rendezvous the code does not perform one
extra, unneeded rendezvous.

No pylint or unit test issues after this change.