Rename runner_lst.py runner_tests.py
Because I've just added runner_updates.py and so runner_lst.py is
no longer a meaningful name: one could ask what list does it refer
to.  Instead, runner_tests is a better name because it clarifies
the purpose of the file.

No breakages in the unit tests added by this commit.

No pylint score changes for the files affected by this commit.