rendezvous/client.py: Use runner_rendezvous.py
Because runner_rendezvous.py does just one thing and is a simpler
module than rendezvous/client.py.  More work to come to reduce the
number or warnings in the latter file and reorganize it further.

No new static errors found by pylint after this commit.

Verified that the new code works like the old one when there are
invalid privacy settings.

Verified that the new code works like the old one when the domain
name is invalid and when the connection is refused.

And, finally, verified that the new code works OK on typical
operating conditions, i.e. when the rendezvous succeeds.