bittorrent/stream.py: forward send_complete to peer.py
Forward send_complete() event to peer.  This is needed by next
patch.  Which will modify peer.py to immediately write more PIECEs
on the socket as soon as send_complete() event is delivered.  In
turn, this is key to put the control of test duration in the hands
of the peer that uploads.