bittorrent/peer.py: allow to control duration at the uploader
Add peer a new self.version attribute.  When this attribute is 1,
the behavior is the traditional one.  Conversely, when the attribute
is 2, test duration is controlled by the uploader.  Since the default
value is 1, this commit introduces no functional change.

When version is 2, the uploader sends a new piece whenever the
send queue is empty.  This allows to control duration, since
it can stop sending as soon as the elapsed upload time is greater
than the target elapsed time.  This form of control is more
robust, because there is no latency caused by queueing in the
equation.