viewer_webkit_gtk.py: add -O setting option
It cannot be -D because that has already a key=value semantic
and I want to use here the same semantic of the config file,
as ssh(1) does with -o option.  I don't use the -o switch tho,
because I think there may be places in Neubot where it makes
sense to save to file, and -o is the natural choice in such
cases.