MacOS/makepkg.py: rmtree: lstat() instead of ignoring errors
I was ignoring errors in rmtree() so that the function call does
not fail if the file/directory to remove does not exist.

Unfortunately ignoring errors is much more strong than -f and has
the collateral effect that also errors in removing an existing tree
are ignored.

I don't want that, so lstat() the target before removing and try
to rmtree it only if it exists.