negotiate: bump default parallelism to 7
This is how the master server was configured (overriding the default
value via command line), I do not see reasons why other servers should
not be configured the same way.  The typical M-Lab site should be on,
at least, a Gigabit, and, to the best of my knowledge, the best you
can get on the world, from a commercial provider, is 100 Mbit/s.