runner_rendezvous.py: repair json-based rendezvous
The server is still unmarshalling into a structure with fixed
names, and the variable name is privacy_can_share.  So, send
a JSON object that contains that name, because if we send the
new "privacy_can_publish" name, the server thinks the client
has not provided enough permissions and will not use M-Lab.