runner_core.py: safely run callback()
This commit wraps callback() invokation with the proper try..except
clauses, to be sure that an error in callback() does not stop the
processing of the runner queue.

While there, s/conf/ctx/, because the object that is passed around
is indeed more a context than configuration.

Finally, if the ctx is not None, pass it to the callback so that
it can use it.

Changes either needed by or inspired by the new Win32 updater,
which has a component that is executed by the runner.

Checked that there are no obvious breakages by running this module
from command line.

No pylint or 2to3 changes.