add updater_win32.py: automatic updates for windows
The toplevel module is updater_win32.py and depends on two
ancillary modules:

updater_runner.py
  This module contains the portable portion of code of
  updater_win32.py, for testability.

updater_utils.py
  Common function that will be shared with MacOS auto
  updater in future.

There are some pylint warnings in updater_runner.py and they
mainly depend TODO and XXX comments in the code.  Indeed, the
current version of this module is not as robust as I would
like, and I have added comments to remember what improvements
are due.

The 2to3 of updater_win32.py is not clean, but the code should
nonetheless work with python3 because I'm distinguishing the
case of Python3 and the case of Python2 when I import the winreg
module of standard library.