net/poller.py: make sure we break loop when no I/O is pending
This has always been the behavior of Neubot poller, and, FWIW it
is also the behavior of asyncore.

After the stdlib-sched patch merge, the behavior has changed and has
become Twisted-like, i.e. the poller waits when there is no pending
I/O instead of exiting from the loop.  This wasn't done by purpose and
was just a side-effect of merging the stdlib-sched patch.

This commit modifies the poller such that it returns to the old
behavior, which is what the rest of Neubot expects.