updater_win32.py: Remember to close open handles before Popen()
Allows to access again the Web User Interface and API after an
automatic update.  Without this flag, the new process inherits the
listening handle, even if it does not know how to access it.
Interestingly, it successfully bind()s to 127.0.0.1:9774, but the
inherited handle "steals" new connections.  So every attempt to
reach 127.0.0.1:9774 eventually triggers a timeout.  The confusing
part is that `netstat -nao` tells you that the inherited handle
is bound by the old PID, not by the new one.