rendezvous/server.py: repair rendezvous server startup
The original code was creating a copy of CONFIG in run() and
then passed it to RendezvousServer.

That was changed, to avoid the copy, but, unfortunately, CONFIG
is not a real dictionary.  (That should be changed.)

To repair, unconditionally make a copy of the passed dictionary
in RendezvousServer, on the basis that it should not modify
global settings.  (And note that this is true, whether CONFIG
is a real dictionary or not.)