bittorrent/peer.py: add code to behave strictly as a bittorrent peer
This commit adds version 3 of the test.  This version differs very
little from version 2, in terms of lines of code.  But the behavior
is radically different.  Indeed, it behaves like a typical bittorrent
peer, and each request is followed by a response.

Despite being a test that sends a piece for each request, version
3 is radically different from version 1 too.  In particular, it
does not attempt to play tricks to fill the pipeline.  As a
consequence, I guess TCP would be buffer limited in most cases.

I don't need this piece of code soon.  Still, I decided to commit
it, because it may be (and, indeed, already is) handy.  There is
value in being able to compare a test that saturates the bottleneck
(version 2) and a test with a strict behavior (version 3).