speedtest/client.py: control test duration at the sender
Add new state variable 'version'.  When version is 2, test
duration is controlled by the sender.  This allows for more
predictable duration, since we avoid all the problems with
network queueing delay that chime in when control is shared
between client and server.

Given the advantages, this diff also initializes to 2 the
version.