speedtest/server.py: control test duration at the sender
When there is no range header, the server will auto-switch to
version 2 of the test.  With version 2, duration is controlled
by the sender.  This allows for more predictable duration,
since we avoid all the problems with network queueing delay
that chime in, instead, when control is shared between client
and server.