regress: neubot/log.py: prune unneeded tests
Now that the logging hack in neubot/log.py is gone, we don't need
anymore to verify that logging.foo() works as expected.  Indeed,
now logging.foo() is foo() in stdlib logging/__init__.py and it can
be safely trusted.

As a bonus, this removes the only pylint warning of this file :).