Globally stop using DNS cache
DNS cache was a counter-measure for when connect() was not
serialized.  It aimed at reducing DNS request overhead of
connect() attempts after the first one.

Since now we (i) have connect() serialization and (ii) we
don't connect() more than one socket at a time, this piece
of code can be thrown into the Attic.