updater_runner updater_utils: channel -> system
I need to specify both the system (`win32`, `macosx`) and the
channel (`testing`, `stable`).  Otherwise I cannot teach to
Win32 updater to follow testing releases.

First step to do that is to rename `system` the `channel`
variable.  Indeed, this variable is used to select the
system-dependent directory on the server.  While the channel
is the name of the file that indicates latest version insi-
de such directory.