http/client.py: compute host header when we know real endpoint
Now it is possible to pass connect() a list of endpoints to
connect() to.  As a consequence, we cannot compute host header
until we know the one in the list that succeeded.  Otherwise,
the URI is invalid because it contains all the addresses in
the list, e.g.::

    http://master.neubot.org master2.neubot.org:9773/rendezvous

As a bonus, the host header now is always computed, regardless
that connect_uri is used.