speedtest/client.py: hack support for backup master server
To start adding more redundancy, I've deployed a backup master
server.  And the rendezvous already is aware of that.  This diff
teaches the speedtest client to look for that server too, when
running on its own context.

This diff is quite hackish because more work is needed to do that
in a more clean way.  But I don't have time to do that right now,
due to pending release and other deadlines concerning the research.