utils_sysdirs.py: rewrite in a more robust way
Written this way, the code is more robust: (i) because we check
for sys.frozen, so we are sure we apply py2exe specific rules
only when we're running under as a frozen executable; (ii) because
each block of code modifies just one variable.

While there, add a main(), such that, in future, it would be
possible to invoke this module directly.