log: stream -> streaming for clarity
The streaming feature deals with passing a client the logs
generated by the neubot daemon while running a test on behalf
of such client.  It has nothing to do with "stream" logging,
where stream here is intended as a stdlib file.

To avoid confusion between stream and streaming, make sure
that names related to streaming feature use the "streaming"
prefix, not the "stream" prefix.