streaming log: filter ACCESS logs in runner client
This is typical UNIX-zen.  The core code is simple and generic
and filtering is implemented at the edges.  Of course, this
is something that may cause performance problems in some cases.
E.g. when you are passing stuff that typically is discarded,
and passing along that stuff causes significant slow downs.

In this respect, sometimes I happen to worry whether this
streaming feature may slow down tests.  But, in any case,
access logs are not something that should increase too much
the load.  Since the bulk of log messages are DEBUGs, and
they were already passed to the client.