updater_runner: also download signature
For now, the signature is not verified, and that will be done
by a subsequent patch.  Note that I needed to add the -O comand
line option, which allows to override the default operating
system, since that was needed to test this code.  Indeed, as
of now, only signatures for MacOS are available on the remote
server `releases.neubot.org`.