updater_win32: implement signature verification
In particular, this is implemented in the portable part of Win32
updater (updater_runner).  Namely, I have moved the code to write
the tarball on disk on the portable part.  And added the call to
OpenSSL right there.

Tested on MacOSX, by running updater_runner against latest MacOS
release, to check whether it is validated.