speedtest results: clarify that it's correlation, not causation
Drop a couple of lines to put it straight that the user broadband
connection is correlated with Neubot results.  There is no causation,
since there are many confounding factors, such as the operating
system buffers, poor wifi connectivity, concurrent users, etc.

While there, make sure that we link the relevant entry of the
frequently asked questions.