Do not trust server-provided update URI
Since the update URI is not signed, we cannot trust it and use
it in the web user interface.  Instead, hardcode our web site
(http://neubot.org), which is what I should have done in the first
place: where do you expect to download it from?  (OK the original
idea was to use the URI to automatically download it, but that
plan failed, when I decided to decouple update and the master
server, so...)

(I was motivated to search for this kind of things after a very
insightful chat with @hellais, thanks!)