utils_path.py: append() result must_be_ascii()
While testing windows autoupdater I noticed that the tarball and
signature path were unicode objects.  Looking at the code, this can
possibly happen only because normalize() returns a unicode.  We've
already sanitized that unicode at the point that it is for sure
ASCII.  Anyway, for correctness, drop an extra must_be_ascii() to
make sure the result is indeed ASCII.  Just because it is confusing
to see unicode strings after utils_path.append().

While there, reorganize the code of append(), such that the tree
of execution is self-evident.