Deal gracefully with database not being writable
I recently removed support for database in the user home directory,
which was confusing because it duplicated user configuration.  Now
this diff attempts to mitigate the fallback, since our tree is full
of places that assume they can write into the database, which of
course now is not always a given.

Specifically, I've added a readonly flag to the database, which is
set when start as an ordinary user.  This is (or should be) also
safe for MacOSX, since the agent is initially started as root and
then it drops privs (resulting in the flag being True).

I've tried to ensure that the most common commands work even when
run by a nonprivileged user, and that they bail our gracefully
instead of printing a confusing stacktrace.  I've not checked all
the secondary commands, tho.  So I guess there'll be some minor
surprises here and there.