updater/unix.py: avoid update check just after startup
This has been typical Neubot behavior, until the diff that
introduced the STATE dictionary changed that.  There can be
technical merit in checking for updates just after startup
rather than 30 minutes after startup, but this amendment
was not intended.  The STATE diff was just wrong, so I am
reverting to the old behavior.

I will consider again this issue after the current release,
and to remind me of that, I'm deploying a comment.