updater_runner.py: note we don't need to fetch the checksum
Deploy a note to remember that we can save the download and check
of the checksum.  Since the digital signature is just the signed
checksum of the file, when the digital signature is good, we know
that the checksum is also good.

Of course, it does no harm to fetch and verify the SH256 as well,
except that the update takes a little bit longer, and that the
code is a little bit more complex than needed.