Allow to invoke runner_rendezvous subcommand and set alternative port
This diff is triggered by the need of helping David Vanes to debug
what happens with its upstream VPN tunnel.  While reasoning about
how to stimulate the tunnel, I noticed it is useful to run arbitrary
neubot modules.  So, I've re-added this possibility to the new
main common module.  Once that is possible, it is quite simple to
amend runner_rendezvous to allow to choose the port.