utils_ctl: read /api/exit nonexistent response
If we read the response, the error happens in the client that is
requesting exit, where it is discared.  On the contrary, if the
client immediately closes the connection, the error may happen on
the server (subject to a race condition) where it will be logged
in red, and may raise bug reports.

We can safely stop returing True/False, since no other piece of code
in Neubot was using it.