Next evolution of networking code
The poller is basically unchanged, except that it was moved from
neubot/net/ to neubot/.  New stream etc. share most code with
old networing code, except that the interface with high level
protocols, e.g. HTTP, has been radically changed.  (In a certain
sense this is a return to January 2011 networking code, given
that now protocol code has to specify a callback for each slow
I/O operation requested, in order to disentangle the tree.)

The plan is to keep around both the new and the old code, to be
able to refine the new code in tree, and to allow for smooth
migration from the old to the new networking code.  I've been
working in background on that for a long time, but nothing beats
having the code in the tree and experimenting with it.