utils_net: don't pull CONFIG
Instead of pulling CONFIG in utils_net, pull it in net/stream.py and
pass the CONFIG['prefer_ipv6'] parameter around.  In terms of modularity,
it was a bad idea to pull something as core as CONFIG in utils_net.py,
since that reduces the chances of reusing it.

For example it's not possible to import utils_net from python3, because
config.py is not python3-compliant.  Instead, with this patch in it is
again possible to import and use utils_net from python3 code.