runner_core: use deferreds to make code more robust
This commit makes background_rendezvous and updater_runner much
more robust, by using deferreds to catch unexpected exceptions
and route in any case the flow, so that events that must always
be scheduled -such as the next check for updates or the next
rendezvous with the master server- are always scheduled.

The simplest solution to do that was to modify the runner_core
to always use deferreds.  As a consequence, I needed to amend
runner_api as well to use deferreds.

While there, remove some extra blank lines in small functions,
harmonize the style of logging a bit, and change the way CONFIG
is accessed for storing new variables.