Bump version number to 0.4.15
Version 0.4.14 was released out of the main branch, in a dedicated
tag (indeed called 0.4.14).  So I need to bump the version number
of the master branch, because the current development version is now
the candidate for 0.4.15.

(Yes, I'm going to release much leass release candidates, now that
I have added code to update the testing channel using the time of
the last modification as an hint.)