Next evolution of http client
This new HTTP client is more compact than the previous one, but
shares a lot of code with it.  The main difference is that now
all the work is done in a single class, i.e. the total
abstraction has been reduced.

As for the new networking code, I am keeping the old code
around, to test and polish the new code in tree, as well as to
allow for smooth transition from the old to the new.