Add the new RAW transmission test
This diff adds a new transmission test, which aims to characterise
the user's broadband connection.  Namely the test measures connect()
time, application level RTT, and download goodput.  During the
download, the test takes snapshots of download goodput, process
CPU usage, and web100 variables.

We are adding this new test to study certain TCP properties that
seem to be interesting for network neutrality.  Another reason
behind the new test is to advance Neubot technology, by leveraging
new networking, HTTP and web100 code.  The idea is to learn from
this experience and refine already existing tests.  Finally, let's
also say that the new test is likely to push ahead the web user
interface as well, since it conveys more information than the
other tests.

As mentioned, more work is due, particularly in the web user
interface.  Still, the code seems to be robust by now and has
passed some robustness tests.  So, I am committing it, such
that it is included in the next snapshot of 0.4.15.

While there, I've also performed the following minor changes:

1. scrap the neubot/www/descr thing and put descriptions into
   neubot/api_results.py, thereby removing one XXX;

2. remove XXX comment from neubot/main/module.py, since the
   comment turned out to be too optimistic.