neubot/http: note the destiny of files inside this directory
Yesterday I merged with the tree the new HTTP client code, so
better to make it clear that the old HTTP interface is now being
deprecated and will eventually be removed.  As happened in the
past, this is not going to happen soon, since I want to perform
a smooth transition from the old to the new code.

Anyway, drop a short note in each file saying what its destiny
is, so that who reads the code is informed.