Networking code: better connect() failure handling
The problem I want to solve is that, if parent.handle_connect()
fails, then someone needs to be notified.  So, the first step is to
make the call to handle_connect() more robust, by using a deferred
object.

But this is not enough, since there are cases when the invoker of
the Handler.connect() is not the Handler itself, but a parent
object.  To deal with this problem, now Handler.connect() returns
the connector instance, which, in turn, allows to register a
callback to be invoked when connect() fails.

This is implemented using a new deferred function, which passes the
argument of callback() to all callbacks in the list, ignoring the
result of each callback.  This is not portable Twisted-like code,
but seems to be very useful for Neubot.

While there, convert neubot/stream.py to use the new non portable
deferred function as well, thereby removing one XXX.