MacOS port moved to github.com/neubot/neubot_macos
I'm moving ports to separate repositories, to allow me to apply
local patches to a port when a release is due.  Speeds up the
release process, because I'm not forced to re-apply all patches
to master and check again all ports.  Too bad that it took me
so much time to figure out that :-).