api/server.py: break out of poller loop the expedite way
If one invokes POLLER.break_loop(), the poller completes the
pending I/O notifications and then exits.

Instead, when we receive the /api/exit request, we want to
break out of the poller loop _immediately_.

So, override the semantic of KeyboardInterrupt, and use it
for breaking the loop the expedite way.