raw test: always transmit web100 snaps, ignore them at client
I currently don't know whether it makes sense to save the wealth
of web100 data on the client side.  For now, transmit the data
back, but don't save it into the database.  (For debugging reasons
this is much better than doing to converse, since I can see the
web100 snap in flight.)