runner_hosts: don't remember previous query results
The static table safety net must kick in each time there is a
discovery failure.  Caching previous results is bad because
(1) when we want back a random host, it breaks randomness, (2)
otherwise, the cached closest host may down, (3) caching the
closest host is actually a way to hide the problem.