networking_code: s/brigade/buff/i to minimize surprise
When you say brigade, someone that is familiar with Apache thinks
to its bucket brigades.  Unfortunately, Neubot brigade was much
simpler than that, therefore leading to surprise when you notice
the (big) difference, especially in terms of flexibility.

The name was initially OK, since I had plans to do something more
similar to Apache bucket brigades.  However, it turns out that the
current networking code for Neubot is OK, and it's now clear that
I dont't want to make Neubot brigade more complex.

So, rename brigade buff to avoid confusion, and propagate changes.